Moves form types inside their directory
